+++ src/main/java/br/com/techchallenge/adapters/controller/donoRestaurante/DonoRestauranteController.java
@@ -2,12 +2,15 @@ package br.com.techchallenge.adapters.controller.donoRestaurante;
 
 import br.com.techchallenge.adapters.converter.DonoRestauranteDTOConverter;
 import br.com.techchallenge.adapters.dto.donoRestaurante.DonoRestauranteRequestDTO;
+import br.com.techchallenge.infra.entity.DonoRestauranteEntity;
 import br.com.techchallenge.infra.service.DonoRestauranteService;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/api/donos-restaurante")
@@ -40,6 +43,16 @@ public class DonoRestauranteController {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Dono de Restaurante não encontrado");
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Erro ao atualizar Dono de Restaurante");
+        }
+    }
+
+    @GetMapping("/buscarTodos")
+    public ResponseEntity<List<DonoRestauranteEntity>> buscarTodos() {
+        try {
+            List<DonoRestauranteEntity> donos = service.buscarTodos();
+            return ResponseEntity.ok(donos);
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
         }
     }
