+++ src/main/java/br/com/techchallenge/domain/donoRestaurante/DonoRestaurante.java
@@ -1,5 +1,6 @@
-package br.com.techchallenge.domain;
+package br.com.techchallenge.domain.donoRestaurante;
 
+import br.com.techchallenge.domain.Endereco;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
